Better handling of xml:base.
If, when loading a catalog, it does not have the xml:base attribute,
we add it with the absolute path of the directory containing the
catalog file.

When resolving a URI, we parse the value as a URL. If it has a
type (or scheme, in the parlance of RFC 3896), we leave it as
is. Otherwise, if it is not an absolute path, we prepend the current
xml:base value.